Adding Power BI translation files  (#4815)
<!-- Thank you for submitting a Pull Request. If you're new to
contributing to BCApps please read our pull request guideline below
* https://github.com/microsoft/BCApps/Contributing.md
-->
#### Summary <!-- Provide a general summary of your changes -->

This PR adds the translation files from the Power BI files here which
were meant to be added in this other PR:
https://github.com/microsoft/BCApps/pull/4757/files

The translation team has a deadline to point the Juno platform to this
repo this week, so we are prioritizing the translation files while the
reminder of the build fails gets sorted out.

#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->
Fixes
[AB#602534](https://dynamicssmb2.visualstudio.com/1fcb79e7-ab07-432a-a3c6-6cf5a88ba4a5/_workitems/edit/602534)

Co-authored-by: Diego Joshua Martínez Pineda <diegojoshuam@microsoft.com>