Clean up app from symbols folder if it already exists before compilation (#4819)
<!-- Thank you for submitting a Pull Request. If you're new to
contributing to BCApps please read our pull request guideline below
* https://github.com/microsoft/BCApps/Contributing.md
-->
#### Summary <!-- Provide a general summary of your changes -->
When we compile apps in clean mode we start by compiling the app in
"default" mode i.e. without preprocessor symbols. However, in some cases
the folder that holds the apps built in "default" mode may already
contain the app we're about to build. To avoid issues, we can remove the
app file before we recompile in default mode to ensure that the .app
file was built without preprocessor symbols etc.

#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->
Related to
[AB#603840](https://dynamicssmb2.visualstudio.com/1fcb79e7-ab07-432a-a3c6-6cf5a88ba4a5/_workitems/edit/603840)