Add gate for duplicate app and object ids  (#4867)
<!-- Thank you for submitting a Pull Request. If you're new to
contributing to BCApps please read our pull request guideline below
* https://github.com/microsoft/BCApps/Contributing.md
-->
#### Summary <!-- Provide a general summary of your changes -->
From time to time we see failures in the uptake due to issues with the
object ids or app ids. Internally, we require that
* App IDs are unique 
* All test objects fall within a certain range 
* Object ids are unique 

This PR adds a gate for that in BCApps. The last-mentioned gate
currently only validates this within the code in BCApps. In the future,
we could look into downloading e.g. the base application source code
prior to running the check so we can also validate that we're not
reusing object ids used in the base application.

#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->
Fixes
[AB#591348](https://dynamicssmb2.visualstudio.com/1fcb79e7-ab07-432a-a3c6-6cf5a88ba4a5/_workitems/edit/591348)