[27.x] Incorrect Billing To Date for Billing Lines (#4621) (#5087)
#### Summary

For a contract starting on **27.02.2020**, billed quarterly with
**"Align to End of Month"** enabled:

This fix corrects the **Billing-To Date** calculation for quarterly
contracts that use the **"Align to End of Month"** option.

- For example, a contract starting on **27.02.2020** with the first
billing period split manually (**27.02.2020–29.02.2020**) and the next
period beginning **01.03.2020** should end on **31.05.2020**.
- Previously, the system incorrectly calculated the end date as
**27.06.2020**.
- The fix ensures that when the start of the billing period is adjusted
to the **first of the month**, the system correctly determines the end
of the full billing quarter, preserving alignment with the “End of
Month” rule.

#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->
Fixes #4401
Fixes

[AB#581883](https://dynamicssmb2.visualstudio.com/1fcb79e7-ab07-432a-a3c6-6cf5a88ba4a5/_workitems/edit/581883)

Co-authored-by: Miljan Milosavljević <miljan83@gmail.com>