[Transaction Storage] Replace dotnet usage with AL implementation. (#5155)
<!-- Thank you for submitting a Pull Request. If you're new to
contributing to BCApps please read our pull request guideline below
* https://github.com/microsoft/BCApps/Contributing.md
-->
#### Summary <!-- Provide a general summary of your changes -->
Transaction storage uses dotnet `Convert` to convert hexadecimal into
decimal.
Replacing this dotnet usage with an AL implementation.

Tested the implementation against the dotnet `Convert.ToInt32(Input,
16)` with a local test.

Test not included in PR as it's a local method and I don't want to make
it internal just to write a test.
Testing the `CalcTenantExportStartTime` method isn't an option as it
requires the aadtenantid to be set and it cannot be controlled.

The testing:
```
    procedure Test()
    var
        alphanumericLbl: Label '0123456789abcdef', Locked = true;
        twochars: Text[2];
        idx1, idx2 : Integer;
    begin
        for idx1 := 1 to StrLen(alphanumericLbl) do
            for idx2 := 1 to StrLen(alphanumericLbl) do begin
                twochars := CopyStr(alphanumericLbl, idx1, 1) + CopyStr(alphanumericLbl, idx2, 1);
                if HexToDecimal(twochars) <> DotNetConvert(twochars) then
                    Error('Mismatch for %1: %2 vs %3', twochars, HexToDecimal(twochars), DotNetConvert(twochars));
            end;
        Message('All tests passed');
    end;

    local procedure DotNetConvert(HexString: Text[2]): Integer
    var
        Convert: DotNet Convert;
    begin
        exit(Convert.ToInt32(HexString, 16));
    end;
```

#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->
Fixes
[AB#605002](https://dynamicssmb2.visualstudio.com/1fcb79e7-ab07-432a-a3c6-6cf5a88ba4a5/_workitems/edit/605002)