Check for necessary information when reading ADI data. (#5161)
<!-- Thank you for submitting a Pull Request. If you're new to
contributing to BCApps please read our pull request guideline below
* https://github.com/microsoft/BCApps/Contributing.md
-->
#### Summary 

When we read data from ADI, in cases where the pdf is not properly read,
or it is a protected pdf, and vendor information is not read, we want to
mark the draft as blank.

That means the user must do the data entry manually.

We reuse the read into buffer impl. we have to quickly parse the output
and check the returned values.

#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->
Fixes
[AB#602142](https://dynamicssmb2.visualstudio.com/1fcb79e7-ab07-432a-a3c6-6cf5a88ba4a5/_workitems/edit/602142)

Co-authored-by: Magnus Hartvig Grønbech <magnushar@microsoft.com>