[27.x] Add AL support for OCV feedback API (#4903) (#5118)
#### Summary <!-- Provide a general summary of your changes --> This PR
adds the following API to allow AL code request submission of feedback
to OCV:
- **RequestFeedback**: Request general feedback for a feature, (or
copilot feature). Also allows specifying an OCV sub-area (featureArea
parameter). Allows the user to select if it is like, dislike, or a
suggestion
- **RequestLikeFeedback**: Requests feedback for a feature, (or copilot
feature) when the user likes the feature. Also allows specifying an OCV
sub-area (featureArea parameter).
**RequestDislikeFeedback**: Requests feedback for a feature, (or copilot
feature) when the user dislikes the feature. Also allows specifying an
OCV sub-area (featureArea parameter).
**SurveyTimerActivity**: OCV provides a concept surrounding usage time,
for example if a user has used a feature for 300 seconds. They provide a
trigger name associated with the survey as the activityName parameter,
and start whether to start or end the timer
**SurveyTriggerActivity**: OCV also provides a non-timer based trigger,
which could be for example, a signal that a user pressed a button. This
takes the activityName parameter for the trigger associated with the OCV
survey.

#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->
Fixes

[AB#609310](https://dynamicssmb2.visualstudio.com/1fcb79e7-ab07-432a-a3c6-6cf5a88ba4a5/_workitems/edit/609310)

---------

<!-- Thank you for submitting a Pull Request. If you're new to
contributing to BCApps please read our pull request guideline below
* https://github.com/microsoft/BCApps/Contributing.md
-->
#### Summary <!-- Provide a general summary of your changes -->

#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->
Fixes #

---------

Co-authored-by: Eddy Lynch <eddylynch@microsoft.com>