[Config Mgt][Test] Add test library for controlling control or treatment allocation in tests (#5218)
<!-- Thank you for submitting a Pull Request. If you're new to
contributing to BCApps please read our pull request guideline below
* https://github.com/microsoft/BCApps/Contributing.md
-->
#### Summary <!-- Provide a general summary of your changes -->
In tests, for now specifically AI tests, we are unable to control the
allocation which makes it difficult to run AI Tests on both control &
treatment.

This gives AI Tests the possibility of controlling which allocation they
want.

#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->
Fixes
[AB#609197](https://dynamicssmb2.visualstudio.com/1fcb79e7-ab07-432a-a3c6-6cf5a88ba4a5/_workitems/edit/609197)

---------

Co-authored-by: Alexander Holstrup <117829001+aholstrup1@users.noreply.github.com>