Log RequiredTestIsolation in build and be specific in the comments (#5301)
#### Summary <!-- Provide a general summary of your changes -->

Reading through the logs, I think the build scripts are correct. Adding
more comments and logs so it's more obvious what's happening.

I believe things are currently failed might be due to side effects of
the tests. When a test has `RequiredTestIsolation = Disabled`, it will
leave traces in the database, which could then cause the follow up tests
to fail

#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->
Fixes #