Adding configuration in PO matching to receive at posting of the invoice (#5303)
# Summary 
Implementation of configuration settings for what to do when getting a
new invoice and the underlying matched order has not yet been marked as
received. There are 5 different configurations a user can select, found
in `EDocPOMConfiguration.Enum.al`:
- "Always ask"
- "Always receive at posting"
- "Never receive at posting"
- "Receive at posting only for certain vendors"
- "Receive at posting except for certain vendors"

Each of these affect the process of processing an invoice in the
following ways:
- When first trying to match
- When deciding whether or not we should warn the user
- When finalizing to create the final invoice.

Additionally an orthogonal setting to allow for receiving lines when
they are of type G/L Account is also considered.

## How the new settings affect the PO matching process

### When first trying to match

When trying to match, we will prevent the match with an error if:
- The configuration is set to "Never receive"
- The configuration is set to "Receive at posting except for certain
vendors", and the vendor of the invoice is selected

This happens in the procedure `MatchPOLinesToEDocumentLine`.

### When deciding whether or not we should warn the user

We want to warn the user if the order has not been received only when
the configuration is set to:
- Always ask
- Never receive
- Receive at posting only for certain vendors (and one of the
non-specified vendors is selected for this invoice)
- Receive at posting except for certain vendors (and one of the
specified vendors is selected for this invoice)

The logic for this is in the procedure `AppendPOMatchWarnings`.

### When finalizing to create the final invoice

We want to ask the user if they want to continue in the following cases:
- Configuration set to "Always ask"
- Configuration set to "Receive at posting only for certain vendors" and
the vendor of the invoice is not one of the configured vendors.

We want to error out in the following cases:
- Configuration set to "Never receive"
- Configuration set to "Receive at posting except for certain vendors"
and one of the selected vendors is the vendor of the invoice

For the rest of the scenarios, we allow finalizing without user
confirmation.

This happens in `EDocCreatePurchaseInvoice.Codeunit.al`.

<img width="2079" height="600" alt="image"
src="https://github.com/user-attachments/assets/39142826-a5d8-4d7c-8407-0172468d08c9"
/>


#### Work Item(s) 
Fixes
[AB#599135](https://dynamicssmb2.visualstudio.com/1fcb79e7-ab07-432a-a3c6-6cf5a88ba4a5/_workitems/edit/599135)

---------

Co-authored-by: Joshua Martínez Pineda <diegojoshuam@microsoft.com>