Revert "[Master]-E-Documents: Check 'Your reference' not available in… (#5422)
… service documents in Belgian localisation (#5347)"

This reverts commit 142bb40712c414daf794943c71b5bafbdb315960.

<!-- Thank you for submitting a Pull Request. If you're new to
contributing to BCApps please read our pull request guideline below
* https://github.com/microsoft/BCApps/Contributing.md
-->
#### Summary <!-- Provide a general summary of your changes -->

Found another issue with BCApp uptake...

The origin PR went in due to: the introduced tests is Uncategorized, so
we only run them in W1 on BCApps.

#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->
Fixes
[AB#611333](https://dynamicssmb2.visualstudio.com/1fcb79e7-ab07-432a-a3c6-6cf5a88ba4a5/_workitems/edit/611333)