[E-Document] Guard against double inserts  (#5448)
#### Summary 

Some scenario the insert could fail due to already existing record. In
such a case we want to avoid failing hard. This PR hardens that by
adding the if then clause around.

#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->
Fixes
[AB#611173](https://dynamicssmb2.visualstudio.com/1fcb79e7-ab07-432a-a3c6-6cf5a88ba4a5/_workitems/edit/611173)

Co-authored-by: Magnus Hartvig Grønbech <magnushar@microsoft.com>