Private/nhsejth/memory telemetry (#5495)
<!-- Thank you for submitting a Pull Request. If you're new to
contributing to BCApps please read our pull request guideline below
* https://github.com/microsoft/BCApps/Contributing.md
-->
#### Summary <!-- Provide a general summary of your changes -->
Add telemetry to the Base64 converter and Json functions that that emits
the source length when then input exceeds a given limit. These functions
are often seen in call stacks when the NST gets OutOfMemory exceptions
and we need to see how large elements are causing this.
There are no functional changes in this PR.

#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->
Fixes
[AB#612984](https://dynamicssmb2.visualstudio.com/1fcb79e7-ab07-432a-a3c6-6cf5a88ba4a5/_workitems/edit/612984)