Extend Price Update in SubscriptionBilling (#5388)
#### Summary 
We want to implement a retroactive price adjustment for the price update
of subscription lines ("Contract Price Update"). To do this we want to
create a copy of a subscription line during the price adjustment and
change the date values of the subscription line depending on the value
provided in the page "Contract Price Update".

Fixes #5374

[AB#613284](https://dynamicssmb2.visualstudio.com/1fcb79e7-ab07-432a-a3c6-6cf5a88ba4a5/_workitems/edit/613284)

---------

Co-authored-by: Jesper Schulz-Wedde <JesperSchulz@users.noreply.github.com>