Update the Base64 Converter app to use the native Baser64Convert (#5651)
<!-- Thank you for submitting a Pull Request. If you're new to
contributing to BCApps please read our pull request guideline below
* https://github.com/microsoft/BCApps/Contributing.md
-->
#### Summary <!-- Provide a general summary of your changes -->
Update the Base64 Converter app to uptake the platform native Base64
converter system codeunit and remove dependencies on DotNet interop
using the Converter class. This is part of the platform push to lower
the risk of out-of-memory exceptions and migrate expensive DotNet
operations to platform code in native codeunits.

#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->
Fixes
[AB#615005](https://dynamicssmb2.visualstudio.com/1fcb79e7-ab07-432a-a3c6-6cf5a88ba4a5/_workitems/edit/615005)