Slice 573373: [World-Class Service] Early Preview Program - Application (#5690)
#### Summary <!-- Provide a general summary of your changes -->
This PR adds a special button and invokes a new event in the Checklist
banner, if the environment is an Early Access Preview environment. In
all other cases, things remain unchanged.

#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->
Fixes
[AB#573373](https://dynamicssmb2.visualstudio.com/1fcb79e7-ab07-432a-a3c6-6cf5a88ba4a5/_workitems/edit/573373)