OpenPasswordChangeDialog allow change password without entering old password (#5814)
<!-- Thank you for submitting a Pull Request. If you're new to
contributing to BCApps please read our pull request guideline below
* https://github.com/microsoft/BCApps/Contributing.md
-->
#### Summary <!-- Provide a general summary of your changes -->

Handle scenario when OpenPasswordChangeDialog is called, but OldPassword
is not entered, so we expect an error about password mismatch.
Add test for that scenario as well.

#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->
Fixes #5587


Fixes
[AB#616472](https://dynamicssmb2.visualstudio.com/1fcb79e7-ab07-432a-a3c6-6cf5a88ba4a5/_workitems/edit/616472)