[Quality Management] - Examples of calling Rec.Modify() in OnValidate triggers #611287 (#5822)
<!-- Thank you for submitting a Pull Request. If you're new to
contributing to BCApps please read our pull request guideline below
* https://github.com/microsoft/BCApps/Contributing.md
-->
#### Summary <!-- Provide a general summary of your changes -->
Calling Modify in the middle of validation code is not very good. We
need to refactor this part of the code.

#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->
Fixes
[AB#611287](https://dynamicssmb2.visualstudio.com/Dynamics%20SMB/_workitems/edit/611287/?view=edit)