[Payables Agent] Improve the demo data implementation of payables agent that is current works on embedded resources (#5743)
### The goal of demo experience for payables agent

The goal of demo experience for payables agent is to have both inbound
e-documents as part of the evaluation companies

<img width="3192" height="713" alt="image"
src="https://github.com/user-attachments/assets/585b2c8a-522e-48f8-b282-c7faf2f6197f"
/>

and also sample invoices will be used when user configures the payables
agent.

<img width="1263" height="558" alt="image"
src="https://github.com/user-attachments/assets/8757c5ae-59f5-4cd6-9d9c-10b0f02603b5"
/>

So, user can run the full scenario of taking the PDF, processing it by
payables agent and having the purchase invoice to post and linked to the
draft document.

### Current implementation

The current implementation use the approach of generating the ADI jsons
and PDFs separately and storing them in the ".resources" folder which
bring the following implementation:
1. This approach will not work with translations
2. Going to fail in the next year because the date is hardcoded in the
json
3. Made it extremely hard to support the Contoso Demotool because it is
required to find the hardcoded value in the json to change it

### New implementation

Contains of three parts:
1. A new way of generating inbound e-documents.
2. Sample purchase invoice functionality.
3. Test codeunit that proves that sample invoices are correctly handled
on top of newly generated inbound e-documents.

_**New way of generating inbound e-documents**_

The new approach is to mock all the required e-document entities. Now it
is going to work for translations, new dates and much easier to make
changes in Contoso Demotool.
A codeunit for reference: **"Contoso Inbound E-Document"**. The goal of
the codeunit is to create purchase invoices, post them and then mock all
the necessary e-document entities to behave like inbound e-documents.

_**Sample purchase invoice functionality**_
The goal of the functionality is to generate PDF and use if for demo
invoices as part of the evaluation company and also for sample invoices
that is used for payables agent.
The core logic is in the **"E-Doc Sample Purchase Invoice"** codeunit
that takes custom purchase header and line tables (these are new tables
with limited set of fields). The Generate method calls the new report,
save the result as PDF and save it in the new **"E-Doc Sample Purch. Inv
File"** table. The record from this table will be used by the Payables
Agent when user run the sample invoice guide in the evaluation company.
Sample invoice functionality is also used by the "E-Document Demodata"
app in the **"Contoso Inbound E-Document"** codeunit.
The reason why the sample invoice functionality is part of the
E-Document Core and not E-Document Demodata is that we have a further
plans on utilizing this functionality for regular tests of payables
agent.

_**Test codeunit**_

Dataset contains a list of sample invoices per country and expected
results after processing.
When the test runs, it takes all the sample invoices from the **"E-Doc
Sample Purch. Inv File"** table per country, processes them through the
ADI, verifies the information in the draft document, finalizees and then
posts the purchase invoice.

_ADO PR where this functionality is used:_

https://dynamicssmb2.visualstudio.com/Dynamics%20SMB/_git/NAV/pullrequest/234414


[AB#579972](https://dynamicssmb2.visualstudio.com/Dynamics%20SMB/_workitems/edit/579972/)