Slice 579972: Demo data implementation of payables agent for new counties. Extra changes. (#5991)
This PR contains minor changes required to adding new demo data for PA:
- Fixing preparation of the temporary buffer before doing historical
matching, specifically for allocation account. The reason is that the
name of the allocation account is different from what we have in the
purchase invoice lines after splitting the allocation account to G/L
accounts.
- Fixing sample purchase invoice functionality by populating the
description on the lines
- Adding new countries to demodata app.json


[AB#579972](https://dynamicssmb2.visualstudio.com/Dynamics%20SMB/_workitems/edit/579972/)